app/requirements.txt
pymongo 
flask 
flask-security
flask-admin
flask-mongoengine
flask_restful
psutil
ipython
requests
PyJWT